Add assertion when double-freeing texture index
Simplify the handling of free texture indices by using `-1` only for indices which are currently in use, whereas the size of the vector is now used to indicate the last free index. Otherwise the assertions incorrectly detect the last texture index always being in use because `-1` was used for both states.

+++ src/engine/client/graphics_threaded.h
@@ -819,7 +819,7 @@ class CGraphics_Threaded : public IEngineGraphics
 	CTextureHandle m_InvalidTexture;
 
 	std::vector<int> m_vTextureIndices;
-	int m_FirstFreeTexture;
+	size_t m_FirstFreeTexture;
 	int m_TextureMemoryUsage;
 
 	std::vector<uint8_t> m_vSpriteHelper;
